+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - **plugin-text**. Added Markdownshortcuts for Lists, Headings and Blockquotes.
+- **core**. Pass plugin props and name of parents to childs too
+- **plugin-text**. Leave nested Text-Plugins on second Enter key press
 
 ## [0.5.0](https://github.com/edtr-io/edtr-io/compare/v0.4.3..v0.5.0) - May 7, 2019